Moved luhn into its own library
