+++ tasks/validate-isin/Cargo.toml
@@ -6,3 +6,6 @@ edition = "2018"
 
 [package.metadata.rosettacode]
 url = "https://rosettacode.org/wiki/Validate_International_Securities_Identification_Number"
+
+[dependencies]
+luhn-test-of-credit-card-numbers = { path = "../luhn-test-of-credit-card-numbers/" }
